Game Modes, Lives, Score Implemented
Now the game have 3 modes: Start, Game, and Game Over. When we are at the Start or Game Over Mode, we just press X to start game.
When we hit the pad with the ball, we get one point, and when the ball hit the bottom of the screen, we loose a life. When we start the game, we have 3 lives and a score of 0.