Double Collisions 99% Fixed
I just needed to add a boolean "collided" to the For Loop that checks collisions with all bricks. Everytime we check a collisions with a brick, we check ALSO if "collided" is false, only then we check the collision. At the very first case of collisions with any brick, "collided" is TRUE, so no other collision will be checked for the same frame. The ball have its deflection managed from the first collision, and no other! This makes that 99% of the time :

- The ball will ONLY collide with one brick at a time, and almost never collide with 2
- The ball almost never change its direction when colliding between 2 bricks, as only one collision happend, and in the good direction

In the now very rare cases when the ball destroys 2 bricks and change it's course, atleast it doesn't do it on the same image, and it's quite easy to track image per image what happend, and so when the ball arrives with chirurgical precision between 2 bricks, it can happend that she penetrate between the 2, and destroy both of them.

Happily from what i can see, i already arrived with a sollution to this last bug:

1.test for collision (ok) | 2.test for deflection (ok) | 3.if true (horizontal defl),test ball_x direction,test brick presence in opposite x direction of hit brick opposed to ball_x direction.if brick presence:differed defl (vertical),else (vertical).test ball_y direction,test brick presence in opposite y direction of hit brick opposed to ball_y direction.if brick presence,differed defl (horizontal) | 5.if brick collision, make a bool true for the rest of the frame,which stops further collision tests,so only 1 collision per frame (ok)

This should fix any (now rare) case where the ball destroys 2 bricks in a very close timing, and change its natural course.